Cleanup admin file
